Initial changes to allow 64bit hid_t values. Removed object library and view from process.

+++ trunk/native/hdf5lib/h5iImp.c
@@ -33,10 +33,10 @@ extern "C" {
 /*
  * Class:     ncsa_hdf_hdf5lib_H5
  * Method:    H5Iget_type
- * Signature: (I)I
+ * Signature: (J)I
  */
 JNIEXPORT jint JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Iget_1type
-  (JNIEnv *env, jclass clss, jint obj_id)
+  (JNIEnv *env, jclass clss, jlong obj_id)
 {
     H5I_type_t retVal = H5I_BADID;
     retVal =  H5Iget_type((hid_t)obj_id);
@@ -56,10 +56,10 @@ JNIEXPORT jint JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Iget_1type
 /*
  * Class:     ncsa_hdf_hdf5lib_H5
  * Method:    H5Iget_name
- * Signature: (ILjava/lang/String;J)J
+ * Signature: (JLjava/lang/String;J)J
  */
 JNIEXPORT jlong JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Iget_1name
-  (JNIEnv *env, jclass clss, jint obj_id, jobjectArray name, jlong buf_size)
+  (JNIEnv *env, jclass clss, jlong obj_id, jobjectArray name, jlong buf_size)
 {
     char *aName;
     jstring str;
@@ -96,10 +96,10 @@ JNIEXPORT jlong JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Iget_1name
 /*
  * Class:     ncsa_hdf_hdf5lib_H5
  * Method:    H5Iget_ref
- * Signature: (I)I
+ * Signature: (J)I
  */
 JNIEXPORT jint JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Iget_1ref
-  (JNIEnv *env, jclass clss, jint obj_id)
+  (JNIEnv *env, jclass clss, jlong obj_id)
 {
     int retVal = -1;
     retVal = H5Iget_ref( (hid_t)obj_id);
@@ -112,10 +112,10 @@ JNIEXPORT jint JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Iget_1ref
 /*
  * Class:     ncsa_hdf_hdf5lib_H5
  * Method:    H5Iinc_ref
- * Signature: (I)I
+ * Signature: (J)I
  */
 JNIEXPORT jint JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Iinc_1ref
-  (JNIEnv *env, jclass clss, jint obj_id)
+  (JNIEnv *env, jclass clss, jlong obj_id)
 {
     int retVal = -1;
     retVal = H5Iinc_ref( (hid_t)obj_id);
@@ -128,10 +128,10 @@ JNIEXPORT jint JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Iinc_1ref
 /*
  * Class:     ncsa_hdf_hdf5lib_H5
  * Method:    H5Idec_1ref
- * Signature: (I)I
+ * Signature: (J)I
  */
 JNIEXPORT jint JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Idec_1ref
-  (JNIEnv *env, jclass clss, jint obj_id)
+  (JNIEnv *env, jclass clss, jlong obj_id)
 {
     int retVal = -1;
     retVal = H5Idec_ref( (hid_t)obj_id);
@@ -152,11 +152,11 @@ JNIEXPORT jint JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Idec_1ref
 /*
  * Class:     ncsa_hdf_hdf5lib_H5
  * Method:    H5Iget_file_id
- * Signature: (I)I
+ * Signature: (J)J
  */
 
-JNIEXPORT jint JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Iget_1file_1id
-  (JNIEnv *env, jclass clss, jint obj_id)
+JNIEXPORT jlong JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Iget_1file_1id
+  (JNIEnv *env, jclass clss, jlong obj_id)
 {
     hid_t file_id = 0;
 
@@ -166,7 +166,7 @@ JNIEXPORT jint JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Iget_1file_1id
         h5libraryError(env);
     }
 
-    return (jint) file_id;
+    return (jlong) file_id;
 }
 
 /**********************************************************************
