+++ trunk/native/hdf5lib/h5lImp.c
@@ -50,10 +50,10 @@ extern "C" {
     /*
      * Class:     ncsa_hdf_hdf5lib_H5
      * Method:    H5Lcopy
-     * Signature: (ILjava/lang/String;ILjava/lang/String;II)V
+     * Signature: (JLjava/lang/String;JLjava/lang/String;JJ)V
      */
     JNIEXPORT void JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Lcopy
-      (JNIEnv *env, jclass clss, jint cur_loc_id, jstring cur_name, jint dst_loc_id, jstring dst_name, jint create_id, jint access_id)
+      (JNIEnv *env, jclass clss, jlong cur_loc_id, jstring cur_name, jlong dst_loc_id, jstring dst_name, jlong create_id, jlong access_id)
     {
         char    *lCurName;
         char    *lDstName;
@@ -100,10 +100,10 @@ extern "C" {
     /*
      * Class:     ncsa_hdf_hdf5lib_H5
      * Method:    H5Lcreate_external
-     * Signature: (Ljava/lang/String;Ljava/lang/String;ILjava/lang/String;II)V
+     * Signature: (Ljava/lang/String;Ljava/lang/String;JLjava/lang/String;JJ)V
      */
     JNIEXPORT void JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Lcreate_1external
-    (JNIEnv *env, jclass clss, jstring file_name, jstring cur_name, jint dst_loc_id, jstring dst_name, jint create_id, jint access_id)
+    (JNIEnv *env, jclass clss, jstring file_name, jstring cur_name, jlong dst_loc_id, jstring dst_name, jlong create_id, jlong access_id)
   {
       char    *lFileName;
       char    *lCurName;
@@ -167,10 +167,10 @@ extern "C" {
     /*
      * Class:     ncsa_hdf_hdf5lib_H5
      * Method:    H5Lcreate_hard
-     * Signature: (ILjava/lang/String;ILjava/lang/String;II)V
+     * Signature: (JLjava/lang/String;JLjava/lang/String;JJ)V
      */
     JNIEXPORT void JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Lcreate_1hard
-      (JNIEnv *env, jclass clss, jint cur_loc_id, jstring cur_name, jint dst_loc_id, jstring dst_name, jint create_id, jint access_id)
+      (JNIEnv *env, jclass clss, jlong cur_loc_id, jstring cur_name, jlong dst_loc_id, jstring dst_name, jlong create_id, jlong access_id)
     {
         char    *lCurName;
         char    *lDstName;
@@ -217,10 +217,10 @@ extern "C" {
     /*
      * Class:     ncsa_hdf_hdf5lib_H5
      * Method:    H5Lcreate_soft
-     * Signature: (Ljava/lang/String;ILjava/lang/String;II)V
+     * Signature: (Ljava/lang/String;JLjava/lang/String;JJ)V
      */
     JNIEXPORT void JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Lcreate_1soft
-      (JNIEnv *env, jclass clss, jstring cur_name, jint dst_loc_id, jstring dst_name, jint create_id, jint access_id)
+      (JNIEnv *env, jclass clss, jstring cur_name, jlong dst_loc_id, jstring dst_name, jlong create_id, jlong access_id)
     {
         char    *lCurName;
         char    *lDstName;
@@ -267,10 +267,10 @@ extern "C" {
     /*
      * Class:     ncsa_hdf_hdf5lib_H5
      * Method:    H5Ldelete
-     * Signature: (ILjava/lang/String;I)V
+     * Signature: (JLjava/lang/String;J)V
      */
     JNIEXPORT void JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Ldelete
-      (JNIEnv *env, jclass clss, jint loc_id, jstring name, jint access_id)
+      (JNIEnv *env, jclass clss, jlong loc_id, jstring name, jlong access_id)
     {
         char    *lName;
         jboolean isCopy;
@@ -300,10 +300,10 @@ extern "C" {
     /*
      * Class:     ncsa_hdf_hdf5lib_H5
      * Method:    H5Ldelete_by_idx
-     * Signature: (ILjava/lang/String;IIJI)V
+     * Signature: (JLjava/lang/String;IIJJ)V
      */
     JNIEXPORT void JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Ldelete_1by_1idx
-    (JNIEnv *env, jclass clss, jint loc_id, jstring name, jint index_field, jint order, jlong link_n, jint access_id)
+    (JNIEnv *env, jclass clss, jlong loc_id, jstring name, jint index_field, jint order, jlong link_n, jlong access_id)
     {
         char      *lName;
         herr_t     status;
@@ -333,10 +333,10 @@ extern "C" {
     /*
      * Class:     ncsa_hdf_hdf5lib_H5
      * Method:    H5Lexists
-     * Signature: (ILjava/lang/String;I)Z
+     * Signature: (JLjava/lang/String;J)Z
      */
     JNIEXPORT jboolean JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Lexists
-      (JNIEnv *env, jclass clss, jint loc_id, jstring name, jint access_id)
+      (JNIEnv *env, jclass clss, jlong loc_id, jstring name, jlong access_id)
     {
         char    *lName;
         jboolean isCopy;
@@ -371,10 +371,10 @@ extern "C" {
     /*
      * Class:     ncsa_hdf_hdf5lib_H5
      * Method:    H5Lget_info
-     * Signature: (ILjava/lang/String;I)Lncsa/hdf/hdf5lib/structs/H5L_info_t;
+     * Signature: (JLjava/lang/String;J)Lncsa/hdf/hdf5lib/structs/H5L_info_t;
      */
     JNIEXPORT jobject JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Lget_1info
-    (JNIEnv *env, jclass clss, jint loc_id, jstring name, jint access_id)
+    (JNIEnv *env, jclass clss, jlong loc_id, jstring name, jlong access_id)
     {
         char      *lName;
         herr_t     status;
@@ -432,10 +432,10 @@ extern "C" {
     /*
      * Class:     ncsa_hdf_hdf5lib_H5
      * Method:    H5Lget_info_by_idx
-     * Signature: (ILjava/lang/String;IIJI)Lncsa/hdf/hdf5lib/structs/H5L_info_t;
+     * Signature: (JLjava/lang/String;IIJJ)Lncsa/hdf/hdf5lib/structs/H5L_info_t;
      */
     JNIEXPORT jobject JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Lget_1info_1by_1idx
-    (JNIEnv *env, jclass clss, jint loc_id, jstring name, jint index_field, jint order, jlong link_n, jint access_id)
+    (JNIEnv *env, jclass clss, jlong loc_id, jstring name, jint index_field, jint order, jlong link_n, jlong access_id)
     {
         char      *lName;
         herr_t     status;
@@ -493,10 +493,10 @@ extern "C" {
     /*
      * Class:     ncsa_hdf_hdf5lib_H5
      * Method:    H5Lget_name_by_idx
-     * Signature: (ILjava/lang/String;IIJI)Ljava/lang/String;
+     * Signature: (JLjava/lang/String;IIJJ)Ljava/lang/String;
      */
     JNIEXPORT jobject JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Lget_1name_1by_1idx
-    (JNIEnv *env, jclass clss, jint loc_id, jstring name, jint index_field, jint order, jlong link_n, jint access_id)
+    (JNIEnv *env, jclass clss, jlong loc_id, jstring name, jint index_field, jint order, jlong link_n, jlong access_id)
     {
         size_t   buf_size;
         char    *lName;
@@ -558,10 +558,10 @@ extern "C" {
     /*
      * Class:     ncsa_hdf_hdf5lib_H5
      * Method:    H5Lget_val
-     * Signature: (ILjava/lang/String;[Ljava/lang/String;I)I
+     * Signature: (JLjava/lang/String;[Ljava/lang/String;J)I
      */
     JNIEXPORT jint JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Lget_1val
-      (JNIEnv *env, jclass clss, jint loc_id, jstring name, jobjectArray link_value, jint access_id)
+      (JNIEnv *env, jclass clss, jlong loc_id, jstring name, jobjectArray link_value, jlong access_id)
     {
         size_t      buf_size;
         herr_t      status;
@@ -660,11 +660,11 @@ extern "C" {
     /*
      * Class:     ncsa_hdf_hdf5lib_H5
      * Method:    H5Lget_val_by_idx
-     * Signature: (ILjava/lang/String;IIJ[Ljava/lang/String;I)I
+     * Signature: (JLjava/lang/String;IIJ[Ljava/lang/String;J)I
      */
     JNIEXPORT jint JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Lget_1val_1by_1idx
-    (JNIEnv *env, jclass clss, jint loc_id, jstring name, jint index_field, jint order, 
-            jlong link_n, jobjectArray link_value, jint access_id)
+    (JNIEnv *env, jclass clss, jlong loc_id, jstring name, jint index_field, jint order,
+            jlong link_n, jobjectArray link_value, jlong access_id)
     {
         herr_t      status;
         size_t      buf_size;
@@ -761,10 +761,10 @@ extern "C" {
     /*
      * Class:     ncsa_hdf_hdf5lib_H5
      * Method:    H5Lmove
-     * Signature: (ILjava/lang/String;ILjava/lang/String;II)V
+     * Signature: (JLjava/lang/String;JLjava/lang/String;JJ)V
      */
     JNIEXPORT void JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Lmove
-      (JNIEnv *env, jclass clss, jint cur_loc_id, jstring cur_name, jint dst_loc_id, jstring dst_name, jint create_id, jint access_id)
+      (JNIEnv *env, jclass clss, jlong cur_loc_id, jstring cur_name, jlong dst_loc_id, jstring dst_name, jlong create_id, jlong access_id)
     {
         char    *lCurName;
         char    *lDstName;
@@ -870,10 +870,10 @@ extern "C" {
     /*
      * Class:     ncsa_hdf_hdf5lib_H5
      * Method:    H5Lvisit
-     * Signature: (IIILjava/lang/Object;Ljava/lang/Object;)I
+     * Signature: (JIILjava/lang/Object;Ljava/lang/Object;)I
      */
     JNIEXPORT jint JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Lvisit
-      (JNIEnv *env, jclass clss, jint grp_id, jint idx_type, jint order,
+      (JNIEnv *env, jclass clss, jlong grp_id, jint idx_type, jint order,
               jobject callback_op, jobject op_data)
     {
         herr_t        status = -1;
@@ -903,11 +903,11 @@ extern "C" {
     /*
      * Class:     ncsa_hdf_hdf5lib_H5
      * Method:    H5Lvisit_by_name
-     * Signature: (ILjava/lang/String;IILjava/lang/Object;Ljava/lang/Object;I)I
+     * Signature: (JLjava/lang/String;IILjava/lang/Object;Ljava/lang/Object;J)I
      */
     JNIEXPORT jint JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Lvisit_1by_1name
-      (JNIEnv *env, jclass clss, jint grp_id, jstring name, jint idx_type, jint order,
-              jobject callback_op, jobject op_data, jint access_id)
+      (JNIEnv *env, jclass clss, jlong grp_id, jstring name, jint idx_type, jint order,
+              jobject callback_op, jobject op_data, jlong access_id)
     {
         jboolean      isCopy;
         char         *lName;
@@ -951,10 +951,10 @@ extern "C" {
     /*
      * Class:     ncsa_hdf_hdf5lib_H5
      * Method:    H5Literate
-     * Signature: (IIIJLjava/lang/Object;Ljava/lang/Object;)I
+     * Signature: (JIIJLjava/lang/Object;Ljava/lang/Object;)I
      */
     JNIEXPORT jint JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Literate
-      (JNIEnv *env, jclass clss, jint grp_id, jint idx_type, jint order,
+      (JNIEnv *env, jclass clss, jlong grp_id, jint idx_type, jint order,
               jlong idx, jobject callback_op, jobject op_data)
     {
         hsize_t       start_idx = (hsize_t)idx;
@@ -985,11 +985,11 @@ extern "C" {
     /*
      * Class:     ncsa_hdf_hdf5lib_H5
      * Method:    H5Literate_by_name
-     * Signature: (ILjava/lang/String;IIJLjava/lang/Object;Ljava/lang/Object;I)I
+     * Signature: (JLjava/lang/String;IIJLjava/lang/Object;Ljava/lang/Object;J)I
      */
     JNIEXPORT jint JNICALL Java_ncsa_hdf_hdf5lib_H5_H5Literate_1by_1name
-      (JNIEnv *env, jclass clss, jint grp_id, jstring name, jint idx_type, jint order,
-              jlong idx, jobject callback_op, jobject op_data, jint access_id)
+      (JNIEnv *env, jclass clss, jlong grp_id, jstring name, jint idx_type, jint order,
+              jlong idx, jobject callback_op, jobject op_data, jlong access_id)
     {
         jboolean      isCopy;
         char         *lName;
